NotificationHolderInterface rename
