+++ automation/src/main/java/info/nightscout/androidaps/services/LocationServiceHelper.kt
@@ -5,7 +5,7 @@ import android.content.Context
 import android.content.Intent
 import android.content.ServiceConnection
 import android.os.IBinder
-import info.nightscout.androidaps.interfaces.NotificationHolderInterface
+import info.nightscout.androidaps.interfaces.NotificationHolder
 import javax.inject.Inject
 import javax.inject.Singleton
 
@@ -20,7 +20,7 @@ import javax.inject.Singleton
  */
 @Singleton
 class LocationServiceHelper @Inject constructor(
-    private val notificationHolder: NotificationHolderInterface
+    private val notificationHolder: NotificationHolder
 ) {
 
     fun startService(context: Context) {
