+++ core/src/main/java/info/nightscout/androidaps/services/AlarmSoundService.kt
@@ -10,7 +10,7 @@ import android.os.IBinder
 import dagger.android.DaggerService
 import info.nightscout.androidaps.activities.ErrorHelperActivity
 import info.nightscout.androidaps.core.R
-import info.nightscout.androidaps.interfaces.NotificationHolderInterface
+import info.nightscout.androidaps.interfaces.NotificationHolder
 import info.nightscout.androidaps.logging.AAPSLogger
 import info.nightscout.androidaps.logging.LTag
 import info.nightscout.androidaps.utils.resources.ResourceHelper
@@ -23,7 +23,7 @@ class AlarmSoundService : DaggerService() {
 
     @Inject lateinit var aapsLogger: AAPSLogger
     @Inject lateinit var resourceHelper: ResourceHelper
-    @Inject lateinit var notificationHolder: NotificationHolderInterface
+    @Inject lateinit var notificationHolder: NotificationHolder
     @Inject lateinit var sp: SP
 
     private var player: MediaPlayer? = null
